+++ libnavui-androidauto/CHANGELOG.md
@@ -4,6 +4,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## Unreleased
 #### Features
+- Added support for MUTCD and Vienna designs in the Speed limit widget. [#6275](https://github.com/mapbox/mapbox-navigation-android/pull/6275)
+
 #### Bug fixes and improvements
 
 ## androidauto-v0.9.0 - Sep 1, 2022